+++ src/mako/validator/rules/Arr.php
@@ -1,0 +1,32 @@
+<?php
+
+/**
+ * @copyright Frederic G. Østby
+ * @license   http://www.makoframework.com/license
+ */
+
+namespace mako\validator\rules;
+
+/**
+ * Arr rule.
+ *
+ * @author Frederic G. Østby
+ */
+class Arr extends Rule implements RuleInterface
+{
+	/**
+	 * {@inheritdoc}
+	 */
+	public function validate($value, array $input): bool
+	{
+		return is_array($value);
+	}
+
+	/**
+	 * {@inheritdoc}
+	 */
+	public function getErrorMessage(string $field): string
+	{
+		return sprintf('The %1$s field must contain an array.', $field);
+	}
+}